marcador dirección v0.2: cambios en el render cmp y ajustes varios

+++ src/game/man/UnitsManager.hpp
@@ -14,8 +14,8 @@ struct UnitsManager {
       void init()                        noexcept;
       void deleteEntity(BECS::entID eid) noexcept;
 
-      void createSoldier(const uint64_t size, const int64_t pos_x, const int64_t pos_y, const Color col) noexcept;
-      void createPlayer( const uint64_t size, const int64_t pos_x, const int64_t pos_y, const Color col) noexcept;
+      void createSoldier(const uint32_t size, const int64_t pos_x, const int64_t pos_y, const Color col) noexcept;
+      void createPlayer( const uint32_t size, const int64_t pos_x, const int64_t pos_y, const Color col) noexcept;
     
       template <typename CMP_t> constexpr std::vector<CMP_t>& getComponentVector()                    noexcept;
       template <typename CMP_t> constexpr CMP_t&              getCmpByEntityID(const BECS::entID eid) noexcept;
